+++ game-frontend/projects/game-frontend/src/app/shared/types/navigation-config.type.ts
@@ -2,4 +2,5 @@ export interface NavigationConfig {
     galaxy: number;
     sector: number;
     quadrant: number;
+    planetId?: number;
 }