+++ game-frontend/projects/game-frontend/src/app/shared/util/planet.util.ts
@@ -1,0 +1,13 @@
+import {Planet} from '@owge/universe';
+import {NavigationConfig} from '../types/navigation-config.type';
+
+export class PlanetUtil {
+    static planetToNavigationConfig(planet: Planet): NavigationConfig {
+        return {
+            galaxy: planet.galaxyId,
+            sector: planet.sector,
+            quadrant: planet.quadrant,
+            planetId: planet.id
+        };
+    }
+}