+++ game-frontend/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 v0.11.0 (latest)
 =================
+* __Improvement:__ [Allow to use url navigation in the galaxy browser #510](https://github.com/KevinGuancheDarias/owge/issues/510)  
 * __Fix:__ [As a player I want to be able to see attack message received during pickupAttack message received while collecting (currently only appears to the attacker collecting). #469](https://github.com/KevinGuancheDarias/owge/issues/469)  
 * __Fix:__ [The stats % given by heroes are not updated if they are defeated (unless a unit is created)- Temporary heroes will update % stats if deleted, but no if not defeated or when activated.](https://github.com/KevinGuancheDarias/owge/issues/468)  
 * __Fix:__ [As admin, I need the ships speed icon to appear in the time specials. #483](https://github.com/KevinGuancheDarias/owge/issues/483)